+++ packages/nouns-contracts/contracts/interfaces/INounsAuctionHouseRewards.sol
@@ -20,7 +20,7 @@ pragma solidity ^0.8.6;
 import { INounsAuctionHouseV2 } from './INounsAuctionHouseV2.sol';
 
 interface INounsAuctionHouseRewards is INounsAuctionHouseV2 {
-    function createBid(uint256 nounId, uint16 clientId) external payable;
+    function createBid(uint256 nounId, uint32 clientId) external payable;
 
-    function biddingClient(uint256 nounId) external returns (uint16 clientId);
+    function biddingClient(uint256 nounId) external returns (uint32 clientId);
 }
